test(e2e): consolidate utils helpers for performance migration
Phase 0 of the rollout plan in docs/proposals/e2e-performance-migration.md.

- Split utils/create-target-channel.ts by concern into channels.ts,
  groups.ts, teams.ts, direct-messages.ts, discussions.ts, messages.ts
  and rooms.ts so specs import by topic instead of a catch-all file.
- Unify sendMessage and sendMessageFromUser into a single
  sendMessage(api, roomId, msg, { threadId?, asUser? }) using
  /chat.sendMessage uniformly.
- Add createThreadReply, inviteUsersToRoom and setRoomTopic helpers
  identified as missing for Phase 2 migrations.
- Re-export the new surface from utils/index.ts so specs stay on the
  './utils' import path.
- Update README helper table to match.
