+++ apps/meteor/tests/e2e/apps/app-modal-interaction.spec.ts
@@ -1,6 +1,6 @@
 import { Users } from '../fixtures/userStates';
 import { HomeChannel } from '../page-objects';
-import { createTargetChannel } from '../utils/create-target-channel';
+import { createTargetChannel } from '../utils';
 import { test, expect } from '../utils/test';
 
 test.use({ storageState: Users.admin.state });
