+++ apps/meteor/tests/e2e/utils/index.ts
@@ -1,6 +1,12 @@
-export * from './create-target-channel';
-export * from './setSettingValueById';
-export * from './getSettingValueById';
+export * from './channels';
+export * from './direct-messages';
+export * from './discussions';
 export * from './getPermissionRoles';
+export * from './getSettingValueById';
+export * from './groups';
+export * from './messages';
+export * from './rooms';
+export * from './setSettingValueById';
 export * from './setUserPreferences';
+export * from './teams';
 export * from './updateOwnUserInfo';
